Don't rely on conversation objects having a `last_message` field.
It seems that in some cases conversation objects retrieved from the
Front API do not have the `last_message` field, resulting in errors when
determining which channel to use when syncing a message back to front.
This change retrieves the last message on a conversation with
a separate API request using the `_links` reference.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>